Move download_file() from peer_review to utils module.
