+++ tests/provider/test_utils.py
@@ -6,9 +6,11 @@ import sys
 from xml.etree.ElementTree import Element, SubElement
 import arrow
 from mock import patch
+from testfixtures import TempDirectory
 from ddt import ddt, data, unpack
 import provider.utils as utils
 import botocore.config
+from tests.activity.classes_mock import FakeResponse
 
 
 @ddt
@@ -388,6 +390,38 @@ class TestContentTypeFromFileName(unittest.TestCase):
     def test_content_type_from_file_name(self, input, expected):
         result = utils.content_type_from_file_name(input)
         self.assertEqual(result, expected)
+
+
+class TestDownloadFile(unittest.TestCase):
+    "tests for download_file()"
+
+    def tearDown(self):
+        TempDirectory.cleanup_all()
+
+    @patch("requests.get")
+    def test_download_file(self, fake_get):
+        "test downloading file by GET request to disk"
+        fake_get.return_value = FakeResponse(200, content=b"test")
+        directory = TempDirectory()
+        from_path = "https://example.org/from.jpg"
+        to_file = os.path.join(directory.path, "to.jpg")
+        user_agent = "test"
+        # invoke
+        result = utils.download_file(from_path, to_file, user_agent)
+        # assert
+        self.assertEqual(result, to_file)
+
+    @patch("requests.get")
+    def test_exception(self, fake_get):
+        "test requests raises exception"
+        fake_get.return_value = FakeResponse(404)
+        directory = TempDirectory()
+        from_path = "https://example.org/from.jpg"
+        to_file = os.path.join(directory.path, "to.jpg")
+        user_agent = "test"
+        # invoke
+        with self.assertRaises(RuntimeError):
+            utils.download_file(from_path, to_file, user_agent)
 
 
 if __name__ == "__main__":
